Изменяет и дополняет логику отображения формы, удаляет лишние компоненты
