--- src/view/type-field-view.js
@@ -1,31 +1,0 @@
-import { createElement } from '../render.js';
-
-function createTypeFieldTemplate() {
-  return `<div class="event__type-wrapper">
-            <label class="event__type  event__type-btn" for="event-type-toggle-1">
-              <span class="visually-hidden">Choose event type</span>
-              <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Event type icon">
-            </label>
-            <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
-            <div class="event__type-list">
-            </div>
-          </div>`;
-}
-
-export default class TypeFieldView {
-  getTemplate() {
-    return createTypeFieldTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
